+++ Source/Runtime/Executor/Mono/MonoInputKeyCodeExecutor.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+namespace LazerLabs.Commands
+{
+    public abstract class MonoInputKeyCodeExecutor : DefaultMonoExecutor
+    {
+        [SerializeField] private KeyCode m_inputKeyCode = default;
+
+        protected virtual void Update()
+        {
+            if (Input.GetKeyUp(m_inputKeyCode))
+            {
+                Execute();
+            }
+        }
+    }
+}