+++ Source/Runtime/Runner/CoroutineCommandRunner.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 namespace LazerLabs.Commands
 {
-    public sealed class CoroutineCommandRunner : ICommandVoid<Func<IEnumerator>>
+    public interface ICoroutineStoppable : ICommand { }
+    public sealed class CoroutineCommandRunner : ICommandVoid<Func<IEnumerator>>, ICoroutineStoppable
     {
         private readonly ICoroutine m_coroutineLogic = default;
         private Coroutine m_coroutine = default;
-
 
         public CoroutineCommandRunner(ICoroutine coroutineLogic)
         {
@@ -17,12 +17,17 @@ namespace LazerLabs.Commands
 
         public void Execute(Func<IEnumerator> v1)
         {
+            Execute();
+            m_coroutine = m_coroutineLogic.StartCoroutine(v1?.Invoke());
+        }
+
+        public void Execute()
+        {
             if (m_coroutine != null)
             {
                 m_coroutineLogic.StopCoroutine(m_coroutine);
                 m_coroutine = null;
             }
-            m_coroutine = m_coroutineLogic.StartCoroutine(v1?.Invoke());
         }
     }
 }